modify get_mask_from_lengths, add torch.Longtensor(lengths)

+++ utils.py
@@ -7,7 +7,7 @@ def get_mask_from_lengths(lengths):
     max_len = torch.max(lengths).item()
     ids = torch.arange(0, max_len, out=torch.LongTensor(max_len))
     # mask = (ids < lengths.unsqueeze(1)).bool()
-    mask = (ids < lengths.unsequeeze(1))
+    mask = (ids < torch.LongTensor(lengths).unsequeeze(1))
     return mask
